fix some bugs in english mode
