+++ src/store/slices/languageSlice.js
@@ -1,182 +1,7 @@
 
 import { createSlice } from "@reduxjs/toolkit";
-
-const englishLanguage = {
-    projectName : "Manage Userslist Project",
-    navItem1 : "Home",
-    navItem2 : "Docs",
-    navItem3 : "About",
-    tableLabel : "Table of names of registered persons",
-    tableTh1 : "ProfilePicture",
-    tableTh2 : "Username",
-    tableTh3 : "IDCode",
-    tableTh4 : "E-mail",
-    tableTh5 : "Date Joining",
-    tableTh6 : "AccessRate",
-    tableTh7 : "Edit User",
-    helpRouteMainHeader : "Docs",
-    helpRouteInstallHeader : "Install project",
-    helpRouteInstallParagraph1 : "To install the project on your system, just enter the following phrase into your Windows terminal:",
-    helpRouteInstallParagraph2 : "To edit and personalize the project, after entering the project path, enter the required packages of the project by entering the following phrase and start personalizing the project!",
-    helpRouteAddUserHeader : "Add a new user",
-    helpRouteAddUserParagraph : "To add a new user, just click on the icon at the bottom left of the page:",
-    helpRouteEditUserHeader : "Edit a user",
-    helpRouteEditUserParagraph1 : "To edit the user, just click on the pencil icon in front of his profile:",
-    helpRouteEditUserParagraph2 : "Clicking on the pencil icon changes the form mode, now you can edit the information. After editing, click on the tick icon.",
-    helpRouteDeleteUserHeader : "Delete a user",
-    helpRouteDeleteUserParahraph : "To delete a user, just click on the trash can icon in front of their profile.",
-    helpRouteShowUserHeader : "View a user's information",
-    helpRouteShowUserParagraph : "To see the full information of a user, just click on his name or profile to display his full profile.",
-    helpRouteEndParagraph : "A place for your documents ...",
-    aboutRouteMainHeader : "About the project",
-    aboutRouteMakersHeader : "Creators",
-    aboutRouteMakersParagraph1 : "The original idea for the project came from Hesam Mousavi, the director of the Roocket website.",
-    aboutRouteMakersParagraph2 : "The UI and UX design of this project, along with its coding, was done by Mohammad Zourmand. The ideas of friends at Rocket Camp, including Mr. Reza Pasalar, have also played a significant role in improving this project.",
-    aboutRouteMakersParagraph3 : "Also, this project is always an open source project and welcomes any changes you request.",
-    aboutRouteAbilitiesHeader : "Capabilities of this project",
-    aboutRouteAbilitiesV3_Header : "Version 3.0",
-    aboutRouteAbilitiesV3_Paraghraph : "Version 3.0 is a version with relatively extensive changes, which is described as follows:",
-    aboutRouteAbilitiesV3_li1 : "Changes in the color scheme of the project UI",
-    aboutRouteAbilitiesV3_li2 : "Add documentation to the project",
-    aboutRouteAbilitiesV3_li3 : "Added a section about the project to the project",
-    aboutRouteAbilitiesV3_li4 : "Added routes to the project to make the project more understandable",
-    aboutRouteAbilitiesV3_li5 : "Added a loading to the project when the data is being loaded from the server.",
-    aboutRouteAbilitiesV2_Header : "Version 2.0",
-    aboutRouteAbilitiesV2_Paraghraph1 : "In version 2.0, the changes are completely on the coding side and are not visible to the user who uses only the program.",
-    aboutRouteAbilitiesV2_Paraghraph2 : "The changes made in version 2.0 are as follows:",
-    aboutRouteAbilitiesV2_li1 : "Added context and Reducer for better state management",
-    aboutRouteAbilitiesV2_li2 : "Storing information in an api using the endapi site created by Roocket Group",
-    aboutRouteAbilitiesV2_li3 : "Summarize and reduce the size of the code",
-    aboutRouteAbilitiesV2_li4 : "Bug fixes",
-    aboutRouteAbilitiesV1_Header : "Version 1.0",
-    aboutRouteAbilitiesV1_Paraghraph : "Version 1.0 or the initial version of the project can do the following:",
-    aboutRouteAbilitiesV1_li1 : "Get information from the user.",
-    aboutRouteAbilitiesV1_li2 : "Display information inside a table to the user!",
-    aboutRouteAbilitiesV1_li3 : "Gives the user the ability to delete or edit previous information.",
-    aboutRouteAbilitiesV1_li4 : "Stores user information in the local storage of the browser to be available next time.",
-    aboutRouteContactUsHeader : "Contact us",
-    aboutRouteContactParagraph : "If you have any suggestions, ideas or criticisms about this project, you can contact us through the following ways:",
-    aboutRouteContactUsli1 : " Telegram : ",
-    aboutRouteContactUsli2 : " GitHub : ",
-    aboutRouteContactUsli3 : " E-mail : ",
-    simpleModalParagraph1 : "Looks like you haven't registered any information yet!",
-    simpleModalParagraph2 : "It's okay! Just click the + button at the bottom of the page!",
-    showUserModalHeader : "User profile",
-    showUserModalParagraph1 : "Name : ",
-    showUserModalParagraph2 : "National Code :",
-    showUserModalParagraph3 : "Email :",
-    showUserModalParagraph4 : "User access rate:",
-    showUserModalParagraph5 : "About user:",
-    showUserModalParagraph6 : "Nothing has been entered yet ...",
-    showUserModalCloseBtn : "Close this window",
-    confirmModalParagraph : "Are you sure you want to delete this person?",
-    confirmModalBtn1 : "Delete it !",
-    confirmModalBtn2 : "No, I regret it !",
-    featuresBtnDarkModeTitle : "Dark Mode",
-    featuresBtnDarkModeli1 : "On",
-    featuresBtnDarkModeli2 : "Off",
-    featuresBtnLanguage : "Language",
-    featuresBtnShowView : "User view mode",
-    featuresBtnShowViewli1 : "Table mode",
-    featuresBtnShowViewli2 : "Lattice mode",
-    addNewUserMainHeader : "New member registration form",
-    addNewUserInput1 : "Name :",
-    addNewUserInput2 : "National Code :",
-    addNewUserInput3 : "Email :",
-    addNewUserInput4 : "User access rate:",
-    addNewUserInput5 : "Admin",
-    addNewUserInput6 : "Simple User",
-    addNewUserBtn : "Add new user",
-    loadingParagraph : "Receiving information from the server ..."
-}
-export const persianLanguage = {
-    projectName : "پروژه مدیریت لیست کاربران",
-    navItem1 : "خانه",
-    navItem2 : "مستندات",
-    navItem3 : "درباره برنامه",
-    tableLabel : "جدول اسامی افراد ثبت نام شده",
-    tableTh1 : "پروفایل",
-    tableTh2 : "نام کاربر",
-    tableTh3 : "کدملی",
-    tableTh4 : "ایمیل",
-    tableTh5 : "تاریخ عضویت",
-    tableTh6 : "دسترسی",
-    tableTh7 : "ویرایش کاربر",
-    helpRouteMainHeader : "مستندات پروژه",
-    helpRouteInstallHeader : "نصب پروژه",
-    helpRouteInstallParagraph1 : "برای نصب پروژه بر روی سیستم خود تنها کافیست عبارت زیر را داخل ترمینال ویندوز خود وارد کنید :",
-    helpRouteInstallParagraph2 : "برای ویرایش و شخصی سازی پروژه نیز پس از ورود به مسیر پروژه، با وارد کردن عبارت زیر پکیج های مورد نیاز پروژه را نصب کنید و شروع به شخصی سازی پروژه نمایید !",
-    helpRouteAddUserHeader : "اضافه کردن یک کاربر جدید",
-    helpRouteAddUserParagraph : "برای اضافه کردن یک کاربر جدید، کافی است روی آیکون پایین سمت چپ صفحه کلیک کنید :",
-    helpRouteEditUserHeader : "ویرایش یک کاربر",
-    helpRouteEditUserParagraph1 : "برای ویرایش کردن کاربر کافیست روی آیکون مداد جلوی مشخصات او کلیک کنید:",
-    helpRouteEditUserParagraph2 : "با کلیک روی آیکون مداد حالت فرم تغییر می کند حال می توانید اطلاعات را ویرایش کنید. پس از ویرایش روی آیکون تیک کلیک کنید.",
-    helpRouteDeleteUserHeader : "حذف یک کاربر",
-    helpRouteDeleteUserParahraph : "برای حذف یک کاربر تنها کافیست روی آیکون سطل زباله جلوی مشخصات او کلیک کنید.",
-    helpRouteShowUserHeader : "مشاهده اطلاعات یک کاربر",
-    helpRouteShowUserParagraph : "برای مشاهده کامل اطلاعات یک کاربر کافیست روی اسم یا مشخصات او کلیک کنید تا مشخصات کامل او به شما نمایش داده شود.",
-    helpRouteEndParagraph : "مکانی برای مستندات شما ...",
-    aboutRouteMainHeader : "درباره پروژه",
-    aboutRouteMakersHeader : "سازندگان",
-    aboutRouteMakersParagraph1 : "ایده اولیه این پروژه از حسام موسوی،  مدیر وبسایت راکت، بوده است.",
-    aboutRouteMakersParagraph2 : "طراحی UI و UX این پروژه، به همراه کدنویسی آن توسط محمد زورمند انجام گردیده است. ایده های دوستان در کمپ ریکت راکت، من جمله آقای رضاپاسالار نیز در بهبود هر چه بیشتر این پروژه نقش بسزایی داشته است.",
-    aboutRouteMakersParagraph3 : "هم چنین این پروژه همیشه یک پروژه منبع باز است و از تغییرات درخواستی شما به شدت استقبال می کند.",
-    aboutRouteAbilitiesHeader : "قابلیت های این پروژه",
-    aboutRouteAbilitiesV3_Header : "نسخه 3.0",
-    aboutRouteAbilitiesV3_Paraghraph : "نسخه 3.0 یک نسخه، با تغییرات نسبتاً گسترده می باشد که شرح آن به صورت زیر است :",
-    aboutRouteAbilitiesV3_li1 : "تغییرات در رنگ بندی UI پروژه",
-    aboutRouteAbilitiesV3_li2 : "اضافه شدن مستندات به پروژه",
-    aboutRouteAbilitiesV3_li3 : "اضافه شدن قسمتی در مورد پروژه، به پروژه",
-    aboutRouteAbilitiesV3_li4 : "اضافه شدن route ها به پروژه برای قابل فهم تر کردن پروژه",
-    aboutRouteAbilitiesV3_li5 : "اضافه شدن یک loading به پروژه برای زمانیکه اطلاعات در حال لود شدن از سمت سرور هستند.",
-    aboutRouteAbilitiesV2_Header : "نسخه 2.0",
-    aboutRouteAbilitiesV2_Paraghraph1 : "در نسخه 2.0 تغییرات صورت گرفته تماماً در سمت کدنویسی بوده و توسط کاربری که تنها از برنامه استفاده می کند قابل مشاهده نمی باشد.",
-    aboutRouteAbilitiesV2_Paraghraph2 : "تغییرات صورت گرفته در نسخه 2.0 به شرح زیر است :",
-    aboutRouteAbilitiesV2_li1 : "اضافه شدن context و  Reducer برای مدیریت بهتر state ها",
-    aboutRouteAbilitiesV2_li2 : "ذخیره شدن اطلاعات در یک api به کمک سایت endapi ساخت شده توسط گروه راکت",
-    aboutRouteAbilitiesV2_li3 : "خلاصه سازی و کم کردن حجم کدها",
-    aboutRouteAbilitiesV2_li4 : "رفع باگ ها",
-    aboutRouteAbilitiesV1_Header : " نسخه 1.0",
-    aboutRouteAbilitiesV1_Paraghraph : "نسخه 1.0 یا همان نسخه اولیه پروژه می تواند موارد زیر را انجام دهد :",
-    aboutRouteAbilitiesV1_li1 : "اطلاعات را از کاربر بگیرد.",
-    aboutRouteAbilitiesV1_li2 : "اطلاعات را داخل یک جدول به کاربر نمایش دهد!",
-    aboutRouteAbilitiesV1_li3 : "به کاربر این توانایی را می دهد که اطلاعات قبلی را حذف یا ویرایش کند.",
-    aboutRouteAbilitiesV1_li4 : "اطلاعات کاربر را در لوکال استوریج مرورگر ذخیره می کند تا در دفعات بعدی در دسترس باشد.",
-    aboutRouteContactUsHeader : "ارتباط با ما",
-    aboutRouteContactParagraph : "چنانچه پیشنهاد، ایده یا انتقادی در مورد این پروژه دارید می توانید از طریق راههای زیر با ما تماس برقرار کنید :",
-    aboutRouteContactUsli1 : "تلگرام :",
-    aboutRouteContactUsli2 : "گیت هاب :",
-    aboutRouteContactUsli3 : "ایمیل :",
-    simpleModalParagraph1 : "به نظر میرسه هنوز اطلاعاتی رو ثبت نکردی !",
-    simpleModalParagraph2 : "اشکالی نداره! فقط کافیه روی دکمه + در پایین صفحه کلیک کنی!",
-    showUserModalHeader : "مشخصات کاربر",
-    showUserModalParagraph1 : "نام و نام خانوادگی :",
-    showUserModalParagraph2 : "کد ملی :",
-    showUserModalParagraph3 : "پست الکترونیکی :",
-    showUserModalParagraph4 : "میزان دسترسی کاربر :",
-    showUserModalParagraph5 : "درباره کاربر :",
-    showUserModalParagraph6 : "هنوز اطلاعاتی وارد نشده است ...",
-    showUserModalCloseBtn : "این پنجره را ببند",
-    confirmModalParagraph : "آیا از حذف اطلاعات این شخص اطمینان دارید ؟",
-    confirmModalBtn1 : "! حذفش کن بره باو",
-    confirmModalBtn2 : "! نه پشیمون شدم",
-    featuresBtnDarkModeTitle : "حالت شب",
-    featuresBtnDarkModeli1 : "فعال",
-    featuresBtnDarkModeli2 : "غیرفعال",
-    featuresBtnLanguage : "زبان برنامه",
-    featuresBtnShowView : "حالت نمایش کاربران",
-    featuresBtnShowViewli1 : "حالت جدول",
-    featuresBtnShowViewli2 : "حالت شبکه ای",
-    addNewUserMainHeader : "فرم ثبت نام اعضای جدید",
-    addNewUserInput1 : "نام و نام خانوادگی :",
-    addNewUserInput2 : "کد ملی :",
-    addNewUserInput3 : "ایمیل :",
-    addNewUserInput4 : "میزان دسترسی کاربر :",
-    addNewUserInput5 : "ادمین",
-    addNewUserInput6 : "عضوساده",
-    addNewUserBtn : "اضافه کردن عضو جدید",
-    loadingParagraph : "در حال دریافت اطلاعات از سرور"
-}
+import { englishLanguage } from "./languageSliceLayout/englishLanguage";
+import { persianLanguage } from "./languageSliceLayout/persianLanguage";
 
 const languageSlice = createSlice({
     name : "language",
