+++ src/store/slices/languageSliceLayout/englishLanguage.js
@@ -1,0 +1,90 @@
+
+
+export const englishLanguage = {
+    projectName : "Manage Userslist Project",
+    navItem1 : "Home",
+    navItem2 : "Docs",
+    navItem3 : "About",
+    tableLabel : "Table of names of registered persons",
+    tableTh1 : "ProfilePicture",
+    tableTh2 : "Username",
+    tableTh3 : "IDCode",
+    tableTh4 : "E-mail",
+    tableTh5 : "Date Joining",
+    tableTh6 : "AccessRate",
+    tableTh7 : "Edit User",
+    helpRouteMainHeader : "Docs",
+    helpRouteInstallHeader : "Install project",
+    helpRouteInstallParagraph1 : "To install the project on your system, just enter the following phrase into your Windows terminal:",
+    helpRouteInstallParagraph2 : "To edit and personalize the project, after entering the project path, enter the required packages of the project by entering the following phrase and start personalizing the project!",
+    helpRouteAddUserHeader : "Add a new user",
+    helpRouteAddUserParagraph : "To add a new user, just click on the icon at the bottom left of the page:",
+    helpRouteEditUserHeader : "Edit a user",
+    helpRouteEditUserParagraph1 : "To edit the user, just click on the pencil icon in front of his profile:",
+    helpRouteEditUserParagraph2 : "Clicking on the pencil icon changes the form mode, now you can edit the information. After editing, click on the tick icon.",
+    helpRouteDeleteUserHeader : "Delete a user",
+    helpRouteDeleteUserParahraph : "To delete a user, just click on the trash can icon in front of their profile.",
+    helpRouteShowUserHeader : "View a user's information",
+    helpRouteShowUserParagraph : "To see the full information of a user, just click on his name or profile to display his full profile.",
+    helpRouteEndParagraph : "A place for your documents ...",
+    aboutRouteMainHeader : "About the project",
+    aboutRouteMakersHeader : "Creators",
+    aboutRouteMakersParagraph1 : "The original idea for the project came from Hesam Mousavi, the director of the Roocket website.",
+    aboutRouteMakersParagraph2 : "The UI and UX design of this project, along with its coding, was done by Mohammad Zourmand. The ideas of friends at Rocket Camp, including Mr. Reza Pasalar, have also played a significant role in improving this project.",
+    aboutRouteMakersParagraph3 : "Also, this project is always an open source project and welcomes any changes you request.",
+    aboutRouteAbilitiesHeader : "Capabilities of this project",
+    aboutRouteAbilitiesV3_Header : "Version 3.0",
+    aboutRouteAbilitiesV3_Paraghraph : "Version 3.0 is a version with relatively extensive changes, which is described as follows:",
+    aboutRouteAbilitiesV3_li1 : "Changes in the color scheme of the project UI",
+    aboutRouteAbilitiesV3_li2 : "Add documentation to the project",
+    aboutRouteAbilitiesV3_li3 : "Added a section about the project to the project",
+    aboutRouteAbilitiesV3_li4 : "Added routes to the project to make the project more understandable",
+    aboutRouteAbilitiesV3_li5 : "Added a loading to the project when the data is being loaded from the server.",
+    aboutRouteAbilitiesV2_Header : "Version 2.0",
+    aboutRouteAbilitiesV2_Paraghraph1 : "In version 2.0, the changes are completely on the coding side and are not visible to the user who uses only the program.",
+    aboutRouteAbilitiesV2_Paraghraph2 : "The changes made in version 2.0 are as follows:",
+    aboutRouteAbilitiesV2_li1 : "Added context and Reducer for better state management",
+    aboutRouteAbilitiesV2_li2 : "Storing information in an api using the endapi site created by Roocket Group",
+    aboutRouteAbilitiesV2_li3 : "Summarize and reduce the size of the code",
+    aboutRouteAbilitiesV2_li4 : "Bug fixes",
+    aboutRouteAbilitiesV1_Header : "Version 1.0",
+    aboutRouteAbilitiesV1_Paraghraph : "Version 1.0 or the initial version of the project can do the following:",
+    aboutRouteAbilitiesV1_li1 : "Get information from the user.",
+    aboutRouteAbilitiesV1_li2 : "Display information inside a table to the user!",
+    aboutRouteAbilitiesV1_li3 : "Gives the user the ability to delete or edit previous information.",
+    aboutRouteAbilitiesV1_li4 : "Stores user information in the local storage of the browser to be available next time.",
+    aboutRouteContactUsHeader : "Contact us",
+    aboutRouteContactParagraph : "If you have any suggestions, ideas or criticisms about this project, you can contact us through the following ways:",
+    aboutRouteContactUsli1 : " Telegram : ",
+    aboutRouteContactUsli2 : " GitHub : ",
+    aboutRouteContactUsli3 : " E-mail : ",
+    simpleModalParagraph1 : "Looks like you haven't registered any information yet!",
+    simpleModalParagraph2 : "It's okay! Just click the + button at the bottom of the page!",
+    showUserModalHeader : "User profile",
+    showUserModalParagraph1 : "Name : ",
+    showUserModalParagraph2 : "National Code :",
+    showUserModalParagraph3 : "Email :",
+    showUserModalParagraph4 : "User access rate:",
+    showUserModalParagraph5 : "About user:",
+    showUserModalParagraph6 : "Nothing has been entered yet ...",
+    showUserModalCloseBtn : "Close this window",
+    confirmModalParagraph : "Are you sure you want to delete this person?",
+    confirmModalBtn1 : "Delete it !",
+    confirmModalBtn2 : "No, I regret it !",
+    featuresBtnDarkModeTitle : "Dark Mode",
+    featuresBtnDarkModeli1 : "On",
+    featuresBtnDarkModeli2 : "Off",
+    featuresBtnLanguage : "Language",
+    featuresBtnShowView : "User view mode",
+    featuresBtnShowViewli1 : "Table mode",
+    featuresBtnShowViewli2 : "Lattice mode",
+    addNewUserMainHeader : "New member registration form",
+    addNewUserInput1 : "Name :",
+    addNewUserInput2 : "National Code :",
+    addNewUserInput3 : "Email :",
+    addNewUserInput4 : "User access rate:",
+    addNewUserInput5 : "Admin",
+    addNewUserInput6 : "Simple User",
+    addNewUserBtn : "Add new user",
+    loadingParagraph : "Receiving information from the server ..."
+}
