+++ src/store/slices/languageSliceLayout/persianLanguage.js
@@ -1,0 +1,90 @@
+
+
+export const persianLanguage = {
+    projectName : "پروژه مدیریت لیست کاربران",
+    navItem1 : "خانه",
+    navItem2 : "مستندات",
+    navItem3 : "درباره برنامه",
+    tableLabel : "جدول اسامی افراد ثبت نام شده",
+    tableTh1 : "پروفایل",
+    tableTh2 : "نام کاربر",
+    tableTh3 : "کدملی",
+    tableTh4 : "ایمیل",
+    tableTh5 : "تاریخ عضویت",
+    tableTh6 : "دسترسی",
+    tableTh7 : "ویرایش کاربر",
+    helpRouteMainHeader : "مستندات پروژه",
+    helpRouteInstallHeader : "نصب پروژه",
+    helpRouteInstallParagraph1 : "برای نصب پروژه بر روی سیستم خود تنها کافیست عبارت زیر را داخل ترمینال ویندوز خود وارد کنید :",
+    helpRouteInstallParagraph2 : "برای ویرایش و شخصی سازی پروژه نیز پس از ورود به مسیر پروژه، با وارد کردن عبارت زیر پکیج های مورد نیاز پروژه را نصب کنید و شروع به شخصی سازی پروژه نمایید !",
+    helpRouteAddUserHeader : "اضافه کردن یک کاربر جدید",
+    helpRouteAddUserParagraph : "برای اضافه کردن یک کاربر جدید، کافی است روی آیکون پایین سمت چپ صفحه کلیک کنید :",
+    helpRouteEditUserHeader : "ویرایش یک کاربر",
+    helpRouteEditUserParagraph1 : "برای ویرایش کردن کاربر کافیست روی آیکون مداد جلوی مشخصات او کلیک کنید:",
+    helpRouteEditUserParagraph2 : "با کلیک روی آیکون مداد حالت فرم تغییر می کند حال می توانید اطلاعات را ویرایش کنید. پس از ویرایش روی آیکون تیک کلیک کنید.",
+    helpRouteDeleteUserHeader : "حذف یک کاربر",
+    helpRouteDeleteUserParahraph : "برای حذف یک کاربر تنها کافیست روی آیکون سطل زباله جلوی مشخصات او کلیک کنید.",
+    helpRouteShowUserHeader : "مشاهده اطلاعات یک کاربر",
+    helpRouteShowUserParagraph : "برای مشاهده کامل اطلاعات یک کاربر کافیست روی اسم یا مشخصات او کلیک کنید تا مشخصات کامل او به شما نمایش داده شود.",
+    helpRouteEndParagraph : "مکانی برای مستندات شما ...",
+    aboutRouteMainHeader : "درباره پروژه",
+    aboutRouteMakersHeader : "سازندگان",
+    aboutRouteMakersParagraph1 : "ایده اولیه این پروژه از حسام موسوی،  مدیر وبسایت راکت، بوده است.",
+    aboutRouteMakersParagraph2 : "طراحی UI و UX این پروژه، به همراه کدنویسی آن توسط محمد زورمند انجام گردیده است. ایده های دوستان در کمپ ریکت راکت، من جمله آقای رضاپاسالار نیز در بهبود هر چه بیشتر این پروژه نقش بسزایی داشته است.",
+    aboutRouteMakersParagraph3 : "هم چنین این پروژه همیشه یک پروژه منبع باز است و از تغییرات درخواستی شما به شدت استقبال می کند.",
+    aboutRouteAbilitiesHeader : "قابلیت های این پروژه",
+    aboutRouteAbilitiesV3_Header : "نسخه 3.0",
+    aboutRouteAbilitiesV3_Paraghraph : "نسخه 3.0 یک نسخه، با تغییرات نسبتاً گسترده می باشد که شرح آن به صورت زیر است :",
+    aboutRouteAbilitiesV3_li1 : "تغییرات در رنگ بندی UI پروژه",
+    aboutRouteAbilitiesV3_li2 : "اضافه شدن مستندات به پروژه",
+    aboutRouteAbilitiesV3_li3 : "اضافه شدن قسمتی در مورد پروژه، به پروژه",
+    aboutRouteAbilitiesV3_li4 : "اضافه شدن route ها به پروژه برای قابل فهم تر کردن پروژه",
+    aboutRouteAbilitiesV3_li5 : "اضافه شدن یک loading به پروژه برای زمانیکه اطلاعات در حال لود شدن از سمت سرور هستند.",
+    aboutRouteAbilitiesV2_Header : "نسخه 2.0",
+    aboutRouteAbilitiesV2_Paraghraph1 : "در نسخه 2.0 تغییرات صورت گرفته تماماً در سمت کدنویسی بوده و توسط کاربری که تنها از برنامه استفاده می کند قابل مشاهده نمی باشد.",
+    aboutRouteAbilitiesV2_Paraghraph2 : "تغییرات صورت گرفته در نسخه 2.0 به شرح زیر است :",
+    aboutRouteAbilitiesV2_li1 : "اضافه شدن context و  Reducer برای مدیریت بهتر state ها",
+    aboutRouteAbilitiesV2_li2 : "ذخیره شدن اطلاعات در یک api به کمک سایت endapi ساخت شده توسط گروه راکت",
+    aboutRouteAbilitiesV2_li3 : "خلاصه سازی و کم کردن حجم کدها",
+    aboutRouteAbilitiesV2_li4 : "رفع باگ ها",
+    aboutRouteAbilitiesV1_Header : " نسخه 1.0",
+    aboutRouteAbilitiesV1_Paraghraph : "نسخه 1.0 یا همان نسخه اولیه پروژه می تواند موارد زیر را انجام دهد :",
+    aboutRouteAbilitiesV1_li1 : "اطلاعات را از کاربر بگیرد.",
+    aboutRouteAbilitiesV1_li2 : "اطلاعات را داخل یک جدول به کاربر نمایش دهد!",
+    aboutRouteAbilitiesV1_li3 : "به کاربر این توانایی را می دهد که اطلاعات قبلی را حذف یا ویرایش کند.",
+    aboutRouteAbilitiesV1_li4 : "اطلاعات کاربر را در لوکال استوریج مرورگر ذخیره می کند تا در دفعات بعدی در دسترس باشد.",
+    aboutRouteContactUsHeader : "ارتباط با ما",
+    aboutRouteContactParagraph : "چنانچه پیشنهاد، ایده یا انتقادی در مورد این پروژه دارید می توانید از طریق راههای زیر با ما تماس برقرار کنید :",
+    aboutRouteContactUsli1 : "تلگرام :",
+    aboutRouteContactUsli2 : "گیت هاب :",
+    aboutRouteContactUsli3 : "ایمیل :",
+    simpleModalParagraph1 : "به نظر میرسه هنوز اطلاعاتی رو ثبت نکردی !",
+    simpleModalParagraph2 : "اشکالی نداره! فقط کافیه روی دکمه + در پایین صفحه کلیک کنی!",
+    showUserModalHeader : "مشخصات کاربر",
+    showUserModalParagraph1 : "نام و نام خانوادگی :",
+    showUserModalParagraph2 : "کد ملی :",
+    showUserModalParagraph3 : "پست الکترونیکی :",
+    showUserModalParagraph4 : "میزان دسترسی کاربر :",
+    showUserModalParagraph5 : "درباره کاربر :",
+    showUserModalParagraph6 : "هنوز اطلاعاتی وارد نشده است ...",
+    showUserModalCloseBtn : "این پنجره را ببند",
+    confirmModalParagraph : "آیا از حذف اطلاعات این شخص اطمینان دارید ؟",
+    confirmModalBtn1 : "! حذفش کن بره باو",
+    confirmModalBtn2 : "! نه پشیمون شدم",
+    featuresBtnDarkModeTitle : "حالت شب",
+    featuresBtnDarkModeli1 : "فعال",
+    featuresBtnDarkModeli2 : "غیرفعال",
+    featuresBtnLanguage : "زبان برنامه",
+    featuresBtnShowView : "حالت نمایش کاربران",
+    featuresBtnShowViewli1 : "حالت جدول",
+    featuresBtnShowViewli2 : "حالت شبکه ای",
+    addNewUserMainHeader : "فرم ثبت نام اعضای جدید",
+    addNewUserInput1 : "نام و نام خانوادگی :",
+    addNewUserInput2 : "کد ملی :",
+    addNewUserInput3 : "ایمیل :",
+    addNewUserInput4 : "میزان دسترسی کاربر :",
+    addNewUserInput5 : "ادمین",
+    addNewUserInput6 : "عضوساده",
+    addNewUserBtn : "اضافه کردن عضو جدید",
+    loadingParagraph : "در حال دریافت اطلاعات از سرور"
+}
